* Fix argument types in lisp_cos(), lisp_sin() and lisp_atan() to avoid     overflows. Closes #1 #2 #3.

+++ src/include/lisp.hpp
@@ -20,9 +20,9 @@ enum { PERM_SPACE,
 
 
 #define FIXED_TRIG_SIZE 360               // 360 degrees stored in table
-extern long sin_table[FIXED_TRIG_SIZE];   // this should be filled in by external module
+extern int32_t sin_table[FIXED_TRIG_SIZE];   // this should be filled in by external module
 #define TBS 1662                          // atan table granularity
-extern unsigned short atan_table[TBS];
+extern uint16_t atan_table[TBS];
 #define NILP(x) (x==NULL)
 #define DEFINEDP(x) (x!=l_undefined)
 class bFILE;
@@ -196,9 +196,9 @@ extern lisp_symbol *lsym_root;
 extern char *space[4],*free_space[4];
 extern int space_size[4];
 void *nth(int num, void *list);
-long lisp_atan2(long dy, long dx);
-long lisp_sin(long x);
-long lisp_cos(long x);
+int32_t lisp_atan2(int32_t dy, int32_t dx);
+int32_t lisp_sin(int32_t x);
+int32_t lisp_cos(int32_t x);
 void restore_heap(void *val, int heap);
 void *mark_heap(int heap);
